# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "FlexPrice API Service",
        "title": "FlexPrice API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
...
        "/invoices": {
            "get": {
                "description": "List invoices with optional filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "List invoices",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "customer_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Subscription ID",
                        "name": "subscription_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Wallet ID",
                        "name": "wallet_id",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Invoice statuses",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start time (RFC3339)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time (RFC3339)",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ListInvoicesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new invoice with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Create a new invoice",
                "parameters": [
                    {
                        "description": "Invoice details",
                        "name": "invoice",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateInvoiceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.InvoiceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/invoices/{id}": {
            "get": {
                "description": "Get detailed information about an invoice",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Get an invoice by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.InvoiceResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/invoices/{id}/finalize": {
            "post": {
                "description": "Finalize a draft invoice",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Finalize an invoice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/invoices/{id}/payment": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update the payment status of an invoice",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Update invoice payment status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payment Status Update Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateInvoicePaymentStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.InvoiceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/invoices/{id}/void": {
            "post": {
                "description": "Void an invoice that hasn't been paid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Void an invoice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateInvoiceLineItemRequest": {
            "type": "object",
            "required": [
                "amount",
                "price_id",
                "quantity"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "metadata": {
                    "$ref": "#/definitions/types.Metadata"
                },
                "meter_id": {
                    "type": "string"
                },
                "period_end": {
                    "type": "string"
                },
                "period_start": {
                    "type": "string"
                },
                "price_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "number"
                }
            }
        },
        "dto.CreateInvoiceRequest": {
            "type": "object",
            "required": [
                "amount_due",
                "currency",
                "customer_id",
                "invoice_type"
            ],
            "properties": {
                "amount_due": {
                    "type": "number"
                },
                "amount_paid": {
                    "type": "number"
                },
                "billing_reason": {
                    "$ref": "#/definitions/types.InvoiceBillingReason"
                },
                "currency": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "invoice_status": {
                    "$ref": "#/definitions/types.InvoiceStatus"
                },
                "invoice_type": {
                    "$ref": "#/definitions/types.InvoiceType"
                },
                "line_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CreateInvoiceLineItemRequest"
                    }
                },
                "metadata": {
                    "$ref": "#/definitions/types.Metadata"
                },
                "payment_status": {
                    "$ref": "#/definitions/types.InvoicePaymentStatus"
                },
                "period_end": {
                    "type": "string"
                },
                "period_start": {
                    "type": "string"
                },
                "subscription_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.InvoiceLineItemResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invoice_id": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/types.Metadata"
                },
                "meter_id": {
                    "type": "string"
                },
                "period_end": {
                    "type": "string"
                },
                "period_start": {
                    "type": "string"
                },
                "price_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "subscription_id": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
        "dto.InvoiceResponse": {
            "type": "object",
            "properties": {
                "amount_due": {
                    "type": "number"
                },
                "amount_paid": {
                    "type": "number"
                },
                "amount_remaining": {
                    "type": "number"
                },
                "billing_reason": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "finalized_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invoice_pdf_url": {
                    "type": "string"
                },
                "invoice_status": {
                    "$ref": "#/definitions/types.InvoiceStatus"
                },
                "invoice_type": {
                    "$ref": "#/definitions/types.InvoiceType"
                },
                "line_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.InvoiceLineItemResponse"
                    }
                },
                "metadata": {
                    "$ref": "#/definitions/types.Metadata"
                },
                "paid_at": {
                    "type": "string"
                },
                "payment_status": {
                    "$ref": "#/definitions/types.InvoicePaymentStatus"
                },
                "period_end": {
                    "type": "string"
                },
                "period_start": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "subscription_id": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                },
                "voided_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.ListInvoicesResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.InvoiceResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateInvoicePaymentStatusRequest": {
            "type": "object",
            "required": [
                "payment_status"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "payment_status": {
                    "$ref": "#/definitions/types.InvoicePaymentStatus"
                }
            }
        },
...
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
...
        "types.InvoiceBillingReason": {
            "type": "string",
            "enum": [
                "SUBSCRIPTION_CREATE",
                "SUBSCRIPTION_CYCLE",
                "SUBSCRIPTION_UPDATE",
                "MANUAL"
            ],
            "x-enum-varnames": [
                "InvoiceBillingReasonSubscriptionCreate",
                "InvoiceBillingReasonSubscriptionCycle",
                "InvoiceBillingReasonSubscriptionUpdate",
                "InvoiceBillingReasonManual"
            ]
        },
...
        "types.InvoicePaymentStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "SUCCEEDED",
                "FAILED"
            ],
            "x-enum-varnames": [
                "InvoicePaymentStatusPending",
                "InvoicePaymentStatusSucceeded",
                "InvoicePaymentStatusFailed"
            ]
        },
        "types.InvoiceStatus": {
            "type": "string",
            "enum": [
                "DRAFT",
                "FINALIZED",
                "VOIDED"
            ],
            "x-enum-varnames": [
                "InvoiceStatusDraft",
                "InvoiceStatusFinalized",
                "InvoiceStatusVoided"
            ]
        },
        "types.InvoiceType": {
            "type": "string",
            "enum": [
                "SUBSCRIPTION",
                "ONE_OFF",
                "CREDIT"
            ],
            "x-enum-varnames": [
                "InvoiceTypeSubscription",
                "InvoiceTypeOneOff",
                "InvoiceTypeCredit"
            ]
        },
        "types.Metadata": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
...
        "v1.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Invalid request payload"
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request payload"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter your API key in the format *x-api-key \u0026lt;api-key\u0026gt;**",
            "type": "apiKey",
            "name": "x-api-key",
            "in": "header"
        }
    }
}